add header files to uip
