+++ networks/uip/map_tcp.cc
@@ -16,13 +16,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "uip.h"
+#include "map_tcp.h"
+#include <arpa/inet.h>
 
 #include "utilities.h"
 #include "logger.h"
 
 
-MapTcpSocket::MapTcpSocket(NetworkBackendInterface* backend, uint32_t sip, uint32_t dip, uint16_t sport, uint16_t dport, int fd)
+MapTcpSocket::MapTcpSocket(Uip* backend, uint32_t sip, uint32_t dip, uint16_t sport, uint16_t dport, int fd)
   : RedirectTcpSocket(backend, sip, dip, sport, dport)
 {
   fd_ = fd;
