+++ networks/uip/meson.build
@@ -1,14 +1,23 @@
 mvisor_sources += files(
-  'core.cc',
   'dhcp.cc',
+  'dhcp.h',
   'icmp.cc',
+  'icmp.h',
   'ipv4.cc',
+  'ipv4.h',
   'map_tcp.cc',
+  'map_tcp.h',
   'redirect_icmp.cc',
+  'redirect_icmp.h',
   'redirect_tcp.cc',
+  'redirect_tcp.h',
   'redirect_udp.cc',
+  'redirect_udp.h',
   'tcp.cc',
+  'tcp.h',
   'udp.cc',
+  'udp.h',
+  'uip.cc',
   'uip.h'
 )
