+++ networks/uip/redirect_icmp.cc
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "uip.h"
+#include "redirect_icmp.h"
 
 #include <fcntl.h>
 #include <arpa/inet.h>
@@ -97,7 +97,7 @@ void RedirectIcmpSocket::StartReading() {
     }
 
     /* FIXME: Limit packet size for Linux driver */
-    auto recv_size = UIP_MAX_UDP_PAYLOAD(packet);
+    auto recv_size = IPV4_MAX_UDP_PAYLOAD(packet);
 
     int ret = recv(fd_, packet->data, recv_size, 0);
     if (ret < 0) {
